Backend/Enterprise-Old-Code/api_services/serializers.py
from rest_framework import serializers
from notifications.models import Notifications
from employees.models import Employee
from projects.models import Projects
from trainings.models import Training


class NotificationsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notifications
        fields = '__all__'


class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = '__all__'


class TrainingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Projects
        fields = '__all__'


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Training
        fields = '__all__'
