Backend/Enterprise-Old-Code/requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.3.4
billiard==3.6.4.0
boto3==1.19.2
botocore==1.22.2
cached-property==1.5.2
celery==5.0.5
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.12
django-allauth==0.45.0
django-cors-headers==3.10.0
django-crispy-forms==1.10.0
django-crontab==0.7.1
django-dajax==0.9.2
django-dajaxice==0.7
django-filter==21.1
django-heroku==0.3.1
django-javascript-settings==1.1.2
django-phonenumber-field==6.0.0
django-storages==1.12.3
djangorestframework==3.12.4
dnspython==2.1.0
ecdsa==0.17.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.4.0
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
mongoengine==0.23.1
oauthlib==3.1.1
phonenumbers==8.12.38
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.19.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
reportlab==3.5.67
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sentry-sdk==1.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
zipp==3.4.1
